requirements.txt
flask
flask-SQLAlchemy
SQLAlchemy
python-dotenv
psycopg2-binary
numpy
moviepy
requests
